升级mybatis p插件

+++ src/main/java/com/xiaoshabao/bill/bd/dto/PaginationAnt.java
@@ -1,28 +1,33 @@
 package com.xiaoshabao.bill.bd.dto;
 
 public class PaginationAnt {
-	
+
 	private long total;
-	private int pageSize;
-	private int current;
-	
+	private long pageSize;
+	private long current;
+
 	public long getTotal() {
 		return total;
 	}
+
 	public void setTotal(long total) {
 		this.total = total;
 	}
-	public int getPageSize() {
+
+	public long getPageSize() {
 		return pageSize;
 	}
-	public void setPageSize(int pageSize) {
+
+	public void setPageSize(long pageSize) {
 		this.pageSize = pageSize;
 	}
-	public int getCurrent() {
+
+	public long getCurrent() {
 		return current;
 	}
-	public void setCurrent(int current) {
+
+	public void setCurrent(long current) {
 		this.current = current;
 	}
-	
+
 }
